requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
black==23.9.1
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.3
Django==4.2.5
django-environ==0.11.2
django-rest-passwordreset==1.3.0
django-rosetta==0.9.9
django-timezone-field==6.0.1
djangorestframework==3.14.0
drf-spectacular==0.26.4
elastic-transport==8.4.1
elasticsearch==8.10.1
flower==2.0.1
hiredis==2.2.3
humanize==4.8.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
pika==1.3.2
platformdirs==3.10.0
polib==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.17
vine==5.0.0
wcwidth==0.2.6
